onchain-lab/pyproject.toml
[tool.poetry]
name = "onchain-lab"
version = "0.1.0"
description = "ONCHAIN LAB ingestion pipelines"
authors = ["ONCHAIN LAB <engineering@onchainlab.io>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.9.0"
pyarrow = "^15.0.0"
typer = { version = "^0.12.3", extras = ["all"] }
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
duckdb = "^0.10.0"
rich = "^13.7.0"
tenacity = "^8.5.0"
tzdata = "^2024.1"
pandas = "^2.2.2"
scikit-learn = "^1.5.2"
xgboost = "^2.1.1"
torch = "^2.4.1"
matplotlib = "^3.9.2"
boruta = "^0.3"
joblib = "^1.4.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-mock = "^3.14.0"
ruff = "^0.6.1"
mypy = "^1.11.0"
mkdocs = "^1.6.0"

[tool.poetry.scripts]
onchain = "src.ingest.cli:app"
price-oracle = "src.price_oracle.cli:app"
onchain-utxo = "src.utxo.cli:app"
onchain-metrics = "src.metrics.cli:app"
onchain-models = "src.models.cli:app"

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S", "TID", "RUF"]
ignore = ["S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
plugins = []
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_optional = true
show_error_codes = true
no_implicit_optional = true
mypy_path = "src"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
